DEVHUB-1813 removing all the references to strapi environmental variables
