+++ .github/workflows/pull-request-checks.yml
@@ -9,8 +9,6 @@ env:
     NPM_EMAIL: ${{ secrets.ARTIFACTORY_NPM_EMAIL }}
     REALM_API_URL: ${{ secrets.REALM_API_URL }}
     REALM_SEARCH_URL: ${{ secrets.REALM_SEARCH_URL }}
-    STRAPI_URL: ${{ secrets.STRAPI_DEV_URL }}
-    STRAPI_API_TOKEN: ${{ secrets.STRAPI_API_TOKEN }}
     NEXTAUTH_URL: http://localhost:3000/developer/api/auth/
     NEXTAUTH_SECRET: ${{ secrets.NEXTAUTH_SECRET }}
     CS_BRANCH: ${{secrets.CS_BRANCH}}
