+++ environments/staging.yaml
@@ -24,8 +24,6 @@ envSecrets:
   REVALIDATE_TOKEN: devcenter-secrets
   SENTRY_AUTH_TOKEN: devcenter-secrets
   SENTRY_DSN: devcenter-secrets
-  STRAPI_URL: devcenter-secrets
-  STRAPI_API_TOKEN: devcenter-secrets
   CS_BRANCH: devcenter-secrets
   CS_DELIVERY_TOKEN: devcenter-secrets
   CS_GRAPHQL_URL: devcenter-secrets
